Restructure ml models so multiple models are possible
Create new parent class for ml model and subclasses will inherit some
basic functions. This makes the code
simplified.
